Various secondary fixes.
Save changes to the interaction fields (title, type).

Fixed bug where the the criterion error was not getting captured.

Fixed bug where the import_row could not be updated because it looks
for duplicates in the validation and find itself.
Now check the id is not same as updated record id.

Deleted XLS file which has bad UTF-8 data. Need to work on CSV only.